git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@8884 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/composite.c
@@ -1615,7 +1615,7 @@ MagickExport MagickBooleanType CompositeImage(Image *image,
               }
               case IntensityCompositeOp:
               {
-                pixel=(double) GetPixelIntensity(composite_image,p);
+                pixel=GetPixelIntensity(composite_image,p);
                 break;
               }
               case LightenIntensityCompositeOp:
